mattress api feature done

+++ Backend/Backend/Repositories/MattressRepository.cs
@@ -1,5 +1,6 @@
-﻿using Backend.Data;
+﻿using Backend.Common.Converters;
 using Backend.DTOs.Mattress;
+using MatCron.Backend.Data;
 using MatCron.Backend.Entities;
 using Microsoft.EntityFrameworkCore;
 
@@ -8,10 +9,11 @@ namespace Backend.Repositories
     public class MattressRepository
     {
         private readonly ApplicationDbContext _context;
-
-        public MattressRepository(ApplicationDbContext context)
+        private readonly OrganisationRepository _organisationRepository;
+        public MattressRepository(ApplicationDbContext context, OrganisationRepository organisationRepository)
         {
             _context = context;
+            _organisationRepository = organisationRepository;
         }
 
         public async Task<IEnumerable<MattressDto>> GetAllMattressesAsync()
@@ -25,7 +27,6 @@ namespace Backend.Repositories
                     BatchNo = m.BatchNo,
                     ProductionDate = m.ProductionDate,
                     MattressTypeId = m.MattressTypeId.ToString(),
-                    GroupId = m.GroupId.ToString(),
                     OrgId = m.OrgId.ToString(),
                     EpcCode = m.EpcCode,
                     Status = m.Status,
@@ -41,18 +42,6 @@ namespace Backend.Repositories
             }
         }
 
-        //public async Task<List<MattressDetailedDto>> GetMattressDetailedAsync()
-        //{
-        //    try
-        //    {
-
-        //    }
-        //    catch (Exception ex)
-        //    {
-        //        Console.WriteLine($"Error in GetMattressSummariesAsync: {ex.Message}");
-        //        throw;
-        //    }
-        //}
 
         public async Task<MattressDto> GetMattressByIdAsync(string id)
         {
@@ -69,7 +58,6 @@ namespace Backend.Repositories
                     BatchNo = result.BatchNo,
                     ProductionDate = result.ProductionDate,
                     MattressTypeId = result.MattressTypeId.ToString(),
-                    GroupId = result.GroupId.ToString(),
                     OrgId = result.OrgId.ToString(),
                     EpcCode = result.EpcCode,
                     Status = result.Status,
@@ -88,22 +76,40 @@ namespace Backend.Repositories
         {
             try
             {
+                MattressType mattressType = await _context.MattressTypes.FindAsync(Guid.Parse(dto.MattressTypeId));
+                if(mattressType == null)
+                {
+                    throw new Exception("Mattress type not found");
+                }
+
+
+                Organisation org = await _context.Organisations.FindAsync(Guid.Parse(dto.OrgId));
+                if (org == null)
+                {
+                    throw new Exception("Organisation not found");
+                }
+
                 Mattress mattress = new Mattress
                 {
                     Uid = Guid.NewGuid(),
-                    BatchNo = dto.BatchNo ?? throw new ,
-                    ProductionDate = dto.ProductionDate,
-                    MattressTypeId = Guid.Parse(dto.MattressTypeId),
-                    GroupId = Guid.Parse(dto.GroupId),
-                    OrgId = Guid.Parse(dto.OrgId),
-                    EpcCode = dto.EpcCode,
-                    Status = dto.Status,
+                    BatchNo = dto.BatchNo ?? throw new Exception("batch number not found"),
+                    ProductionDate = dto.ProductionDate != null ? dto.ProductionDate : throw new Exception("Production date not found"),
+                    MattressTypeId = dto.MattressTypeId != null? Guid.Parse(dto.MattressTypeId): throw new Exception("mattress type id not found"),
+                    OrgId = dto.OrgId != null ? Guid.Parse(dto.OrgId) : throw new Exception("org id not found"),
+                    EpcCode = dto.EpcCode ?? "",
+                    Status = dto.Status ?? 1,
                     LifeCyclesEnd = dto.LifeCyclesEnd,
-                    DaysToRotate = dto.DaysToRotate
+                    DaysToRotate = dto.DaysToRotate ?? 0
                 };
                 _context.Mattresses.Add(mattress);
                 await _context.SaveChangesAsync();
-                return dto;
+
+                mattressType.Stock++;
+                _context.MattressTypes.Update(mattressType);
+                await _context.SaveChangesAsync();
+
+                MattressDto result = MattressConverter.ConvertToDto(mattress);
+                return result;
             }
             catch (Exception ex)
             {
@@ -121,14 +127,28 @@ namespace Backend.Repositories
                 {
                     throw new Exception("Mattress not found");
                 }
-                result.Status = dto.Status != null ? dto.Status : result.Status;
+                result.Status = dto.Status?? result.Status;
                 result.LifeCyclesEnd = dto.LifeCyclesEnd ?? result.LifeCyclesEnd;
-                result.DaysToRotate = dto.DaysToRotate != null ? dto.DaysToRotate : result.DaysToRotate;
+                result.DaysToRotate = dto.DaysToRotate ?? result.DaysToRotate;
+
+                if (dto.MattressTypeId != null)
+                {
+                    MattressType type = await _context.MattressTypes.FindAsync(Guid.Parse(dto.MattressTypeId));
+                    if (type == null)
+                    {
+                        throw new Exception("Mattress type not found");
+                    }
+                    result.MattressTypeId = Guid.Parse(dto.MattressTypeId);
+                }
 
 
-                _context.Entry(result).CurrentValues.SetValues(dto);
+                _context.Mattresses.Update(result);
                 await _context.SaveChangesAsync();
 
+
+
+
+                return MattressConverter.ConvertToDto(result);
 
             }
             
@@ -144,12 +164,26 @@ namespace Backend.Repositories
             try
             {
                 Mattress result = await _context.Mattresses.FindAsync(Guid.Parse(id));
+                
                 if (result == null)
                 {
                     return "Mattress not found";
                 }
+
+                MattressType mattressType = await _context.MattressTypes.FindAsync(result.MattressTypeId);
+                if (mattressType == null)
+                {
+                    throw new Exception("Mattress type not found");
+                }
+
+                mattressType.Stock--;
+                _context.MattressTypes.Update(mattressType);
+                await _context.SaveChangesAsync();
+
                 _context.Mattresses.Remove(result);
                 await _context.SaveChangesAsync();
+
+                                
                 return "Mattress deleted successfully";
             }
             catch (Exception ex)
